backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
psycopg[binary]==3.2.9
python-dotenv==1.2.2
